+++ RUFAS/biophysical/manure/processor_enum.py
@@ -9,6 +9,7 @@ from RUFAS.biophysical.manure.separator.separator import Separator
 from RUFAS.biophysical.manure.storage.anaerobic_lagoon import AnaerobicLagoon
 from RUFAS.biophysical.manure.storage.bedded_pack import BeddedPack
 from RUFAS.biophysical.manure.storage.composting import Composting
+from RUFAS.biophysical.manure.storage.daily_spread import DailySpread
 from RUFAS.biophysical.manure.storage.open_lot import OpenLot
 from RUFAS.biophysical.manure.storage.slurry_storage_outdoor import SlurryStorageOutdoor
 from RUFAS.biophysical.manure.storage.slurry_storage_underfloor import SlurryStorageUnderfloor
@@ -40,6 +41,8 @@ class ProcessorType(Enum):
     Composting = Composting
     BeddedPack = BeddedPack
     OpenLot = OpenLot
+
+    DailySpread = DailySpread
 
     @classmethod
     def get_processor_class(cls, processor_type: str) -> Type["Processor"]: