+++ changelog.md
@@ -229,8 +229,8 @@ v0.9.2
 - [2552](https://github.com/RuminantFarmSystems/MASM/pull/2552) - [minor change] Reduce example run warning count.
 - [2551](https://github.com/RuminantFarmSystems/MASM/pull/2551) - [minor change] Raises error in main.py to direct user to errors file for debugging.
 - [2547](https://github.com/RuminantFarmSystems/MASM/pull/2547) - [minor change] [Input Manager] 1/5 Cross Validation basic setup.
+- [2559](https://github.com/RuminantFarmSystems/MASM/pull/2559) - [minor change] [Manure] Implement daily spread processor.
 - [2550](https://github.com/RuminantFarmSystems/MASM/pull/2550) - [minor change] Create metadata for no animal scenario.
-
 
 ### v0.9.2